style: move to router directory
