+++ src/router/index.ts
@@ -1,0 +1,2 @@
+export * from './router';
+export * from './routes';
